# docs/openapi/moodtune.yaml
openapi: 3.0.3
info:
  title: MoodTune API
  version: "1.0.0"
  description: |
    API REST del proyecto **MoodTune**. Provee servicios para:
    - Gestión de usuarios y sesiones
    - Registro de inferencias (FER)
    - Generación y consulta de playlists
    - Reglas de mapeo emoción→parámetros musicales
    - Auditoría de acciones
    - Gestión segura de tokens OAuth de proveedores musicales

    Arquitectura objetivo: núcleo relacional (RDBMS) + perímetro efímero BaaS/objeto.
  termsOfService: https://moodtune.com/terms
  contact:
    name: Equipo 26 — MoodTune
    url: https://moodtune.com/moodtune
    email: equipo26@moodtune.com
servers:
  - url: https://api.moodtune.com/v1
    description: Entorno de Producción
  - url: http://localhost:8000
    description: Entorno local de desarrollo

# Seguridad global: todas las operaciones requieren Bearer JWT,
# salvo aquellas que hagan override con security: []
security:
  - BearerAuth: []
...
paths:
...
  /oauth-tokens:
    post:
      tags: [OAuthTokens]
      summary: Crear o reemplazar token OAuth (upsert por user_id + provider)
      operationId: upsertOAuthToken
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/OAuthTokenUpsert" } } }
      responses:
        "201": { description: Token creado/actualizado, content: { application/json: { schema: { $ref: "#/components/schemas/OAuthToken" } } } }
        "400": { description: Petición inválida, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: No autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Prohibido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      tags: [OAuthTokens]
      summary: Listar tokens
      operationId: listOAuthTokens
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: user_id
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: provider
          in: query
          required: false
          schema: { $ref: "#/components/schemas/MusicProvider" }
      responses:
        "200": { description: Lista paginada, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedOAuthTokens" } } } }
        "401": { description: No autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Prohibido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /oauth-tokens/{tokenId}:
    get:
      tags: [OAuthTokens]
      summary: Obtener token por ID
      operationId: getOAuthToken
      parameters: [ { $ref: "#/components/parameters/TokenId" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OAuthToken" } } } }
        "404": { description: No encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: No autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Prohibido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    patch:
      tags: [OAuthTokens]
      summary: Actualizar parcialmente token
      operationId: patchOAuthToken
      parameters: [ { $ref: "#/components/parameters/TokenId" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/OAuthTokenUpdate" } } }
      responses:
        "200": { description: Actualizado, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateAck" } } } }
        "404": { description: No encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: No autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Prohibido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      tags: [OAuthTokens]
      summary: Eliminar token
      operationId: deleteOAuthToken
      parameters: [ { $ref: "#/components/parameters/TokenId" } ]
      responses:
        "200": { description: Eliminado, content: { application/json: { schema: { $ref: "#/components/schemas/DeleteAck" } } } }
        "404": { description: No encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: No autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Prohibido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /oauth-tokens/valid/{userId}/{provider}:
    get:
      tags: [OAuthTokens]
      summary: Obtener token válido (con refresco automático)
      operationId: getValidOAuthToken
      description: |
        Obtiene un token de acceso válido para el usuario y proveedor especificado.
        Si el token está expirado (con buffer de 5 minutos), lo refresca automáticamente
        usando el servicio de música y actualiza la base de datos.
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: Identificador de usuario
        - name: provider
          in: path
          required: true
          schema: { $ref: "#/components/schemas/MusicProvider" }
          description: Proveedor musical
      responses:
        "200":
          description: Token válido (posiblemente refrescado)
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider: { $ref: "#/components/schemas/MusicProvider" }
                  access_token: { type: string, description: Token de acceso descifrado }
                  expires_at: { type: string, format: date-time }
                  was_refreshed: { type: boolean, description: Indica si el token fue refrescado }
                required: [provider, access_token, expires_at, was_refreshed]
        "404": { description: Token no encontrado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "500": { description: Token corrupto en BD, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "501": { description: Refresh no implementado para este proveedor, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { description: No se pudo refrescar el token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: No autenticado, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Prohibido, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      description: Tamaño de página (1-100)
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      description: Desplazamiento inicial (>=0)
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
...
    TokenId:
      name: tokenId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Identificador de token
  schemas:
...
    MusicProvider:
      type: string
      description: Proveedor musical
      enum: [spotify, apple_music, amazon_music]
    Error:
      type: object
      properties:
        error: { type: string, description: Descripción del error }
        detail: { type: string, description: Detalle técnico adicional }
      required: [error]
    UpdateAck:
      type: object
      properties:
        updated: { type: boolean }
        id: { type: string, description: "ID afectado (si aplica)" }
    DeleteAck:
      type: object
      properties:
        deleted: { type: boolean }
        id: { type: string, description: "ID eliminado (si aplica)" }
...
    OAuthTokenBase:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/MusicProvider" }
        expires_at: { type: string, format: date-time, nullable: true }
    OAuthTokenUpsert:
      allOf:
        - $ref: "#/components/schemas/OAuthTokenBase"
        - type: object
          properties:
            token_id: { type: string, format: uuid }
            access_cipher_b64: { type: string, description: Ciphertext en Base64 (cifrado en app) }
            refresh_cipher_b64: { type: string, description: Ciphertext en Base64 (cifrado en app), nullable: true }
          required: [user_id, provider, access_cipher_b64]
    OAuthTokenUpdate:
      type: object
      properties:
        provider: { $ref: "#/components/schemas/MusicProvider" }
        access_cipher_b64: { type: string }
        refresh_cipher_b64: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
    OAuthToken:
      allOf:
        - $ref: "#/components/schemas/OAuthTokenBase"
        - type: object
          properties:
            token_id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
          required: [token_id, created_at, updated_at]
    PaginatedOAuthTokens:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/OAuthToken" } }
        limit: { type: integer }
        offset: { type: integer }